+++ Cargo.toml
@@ -5,9 +5,17 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
+[lib]
+path = "src/lib.rs"
+
+[[bin]]
+path = "src/main.rs"
+name = "godsaeng-backend"
+
 [dependencies]
 actix-web = "4.2.1"
 chrono = { version = "0.4.23", default-features = false }
 dotenv = "0.15.0"
 serde = { version = "1.0.152", features = ["derive"] }
 sqlx = { version = "0.6.2", features = ["runtime-async-std-native-tls", "postgres", "chrono"] }
+tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
