+++ src/lib.rs
@@ -1,0 +1,8 @@
+pub mod routes;
+pub mod runner;
+
+use sqlx::{Pool, Postgres};
+
+pub struct AppState {
+    db: Pool<Postgres>,
+}
